refactor(scroll-stories): reorganize EEI story components and modules for more scroll stories
