--- src/scripts/components/stories/story/blocks/story-eei/modules/kettle-amount/year-slider/year-slider.tsx
@@ -1,40 +1,0 @@
-import { motion, useTransform } from "motion/react";
-import { useScrollModule } from "../../base-scroll/use-scroll-module";
-import { KettleAmountAnimationConfig } from "../kettle-amount";
-
-import styles from "./year-slider.module.css";
-
-export default function YearSlider() {
-  const { config, scrollYProgress } =
-    useScrollModule<KettleAmountAnimationConfig>();
-  const startYear = 1991;
-  const numberOfYears = 35;
-
-  return (
-    <motion.div
-      className={styles.yearSlider}
-      style={{
-        scale: useTransform(
-          scrollYProgress,
-          config.yearSlider.fadeIn.input,
-          config.yearSlider.fadeIn.output,
-        ),
-      }}
-    >
-      <motion.div
-        className={styles.container}
-        style={{
-          x: useTransform(
-            scrollYProgress,
-            config.yearSlider.slide.input,
-            config.yearSlider.slide.output,
-          ),
-        }}
-      >
-        {Array.from({ length: numberOfYears }).map((_, index) => (
-          <span key={index}>{startYear + index}</span>
-        ))}
-      </motion.div>
-    </motion.div>
-  );
-}
