+++ src/scripts/components/stories/story/scroll-story/modules/animated-arrows/arrows/arrows.module.css
