+++ src/scripts/components/stories/story/scroll-story/story-eei/constants/globe.ts
