+++ src/scripts/components/stories/story/scroll-story/story-eei/globe-scroll.tsx
