+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/bulb-animation/bulb-animation.module.css
