+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/kettle-amount.tsx
@@ -1,0 +1,114 @@
+import ScrollModule from "../../../modules/base-scroll/module/scroll-module";
+import BulbAnimation from "./bulb-animation/bulb-animation";
+import KettleBox from "./kettle-box/kettle-box";
+import KettleOverlay, {
+  KettleOverlayProps,
+} from "./kettle-overlay/kettle-overlay";
+
+import { useModuleContent } from "../../../../../../../providers/story/module-content/use-module-content";
+import { StoryEEIModule } from "../../../../../../../types/story";
+
+import ScrollText from "../../../modules/base-scroll/scroll-text/scroll-text";
+
+const animationConfig = {
+  initial: {
+    scale: {
+      input: [0, 0.04, 0.05, 0.17, 0.3, 0.95, 1],
+      output: [0.01, 0.01, 0.4, 0.4, 1, 1, 0.01],
+    },
+    yPosition: {
+      input: [0, 0.04, 0.16, 0.2],
+      output: ["-40px", "200px", "200px", "0px"],
+    },
+    opacity: {
+      input: [0, 0.98, 1],
+      output: ["1", "1", "0"],
+    },
+    bulbOpacity: {
+      input: [0, 0.09, 0.12],
+      output: ["0", "0", "1"],
+    },
+    text1: {
+      input: [0, 0.065, 0.09],
+      output: ["1", "1", "0"],
+    },
+    text2: {
+      input: [0, 0.09, 0.12],
+      output: ["0", "0", "1"],
+    },
+  },
+  bulbExit: {
+    input: [0.16, 0.2],
+    output: ["0vh", "-100vh"],
+  },
+  squareMeterScale: {
+    input: [0.15, 0.3, 0.5],
+    output: ["6", "6", "1"],
+  },
+  kettleRows: {
+    entryRangeStart: 0.35,
+    entryRangeEnd: 0.55,
+    exitRangeStart: 0.87,
+    exitRangeEnd: 0.89,
+    opacity: {
+      input: [0, 0.89, 0.9],
+      output: ["1", "1", "0"],
+    },
+  },
+  yearSlider: {
+    fadeIn: {
+      input: [0, 0.4, 0.45, 0.89, 0.9],
+      output: ["0", "0", "1", "1", "0"],
+    },
+    slide: {
+      input: [0, 0.84, 0.87],
+      output: ["-49%", "-49%", "37.5%"],
+    },
+  },
+  scrollText: {
+    input: [0, 0.87, 0.95, 0.98, 1],
+    output: ["100%", "100%", "0%", "0%", "-100%"],
+  },
+};
+
+export type KettleAmountAnimationConfig = typeof animationConfig;
+
+export default function KettleAmountModule() {
+  const { module, getRefCallback } = useModuleContent();
+  const eeiModule = module as StoryEEIModule;
+
+  const overlays: Array<KettleOverlayProps> = [
+    {
+      inputRange: [0.55, 0.57, 0.62, 0.64],
+      text: eeiModule.content?.overlay1 || "",
+    },
+    {
+      inputRange: [0.7, 0.77, 0.82, 0.84],
+      text: eeiModule.content?.overlay2 || "",
+    },
+    {
+      inputRange: [0.85, 0.87, 0.88, 0.89],
+      text: eeiModule.content?.overlay3 || "",
+    },
+  ];
+
+  return (
+    <ScrollModule
+      config={animationConfig}
+      lengthFactor={eeiModule.lengthFactor}
+    >
+      <ScrollModule.StickyContainer isGrid ref={getRefCallback(0, 0)}>
+        <BulbAnimation />
+        <KettleBox />
+        {overlays.map((overlay, index) => (
+          <KettleOverlay key={index} {...overlay} />
+        ))}
+        <ScrollText
+          text={eeiModule.content?.scrollText || ""}
+          inputRange={animationConfig.scrollText.input}
+          outputRange={animationConfig.scrollText.output}
+        />
+      </ScrollModule.StickyContainer>
+    </ScrollModule>
+  );
+}
