+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/kettle-box/kettle-box.module.css
