+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/kettle-overlay/kettle-overlay.module.css
