+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/square-meter-label/square-meter-label.tsx
@@ -1,0 +1,25 @@
+import { motion, useTransform } from "motion/react";
+import { useScrollModule } from "../../../../modules/base-scroll/use-scroll-module";
+import { KettleAmountAnimationConfig } from "../kettle-amount";
+
+import styles from "./square-meter-label.module.css";
+
+export default function SquareMeterLabel() {
+  const { scrollYProgress, config } =
+    useScrollModule<KettleAmountAnimationConfig>();
+
+  return (
+    <motion.span
+      className={styles.squareMeter}
+      style={{
+        scale: useTransform(
+          scrollYProgress,
+          config.squareMeterScale.input,
+          config.squareMeterScale.output,
+        ),
+      }}
+    >
+      1m²
+    </motion.span>
+  );
+}
