+++ src/scripts/components/stories/story/scroll-story/story-eei/modules/kettle-amount/year-slider/year-slider.module.css
