+++ src/scripts/components/stories/story/scroll-story/story-eei/satellite-animation/satellite-animation.tsx
